+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 # CHANGELOG
 
-## 0.1.0 (unreleased)
+## 0.3.0 (unreleased)
+
+ToDos:
+ - simplify supplier handling
+ 
+Fixes:
+
+ - implementing hashCode() & equals()
+
+Improvements:
+
+ - added more Unit tests
+
+## 0.2.0
+
+Fixes:
+
+ - CI builds
+
+## 0.1.0
 
 Features: